tease apart backend and models code into submodules
