+++ src/backend/make_history.rs
@@ -1,0 +1,3 @@
+  fn make_history(corpus_id: i32, service_id: i32, owner: &str, description: &str) {
+
+  }
