+++ src/models/history.rs
@@ -1,9 +1,9 @@
-use diesel::pg::PgConnection;
-use diesel::result::Error;
-use diesel::update;
-use diesel::*;
+// use diesel::pg::PgConnection;
+// use diesel::result::Error;
+// use diesel::update;
+// use diesel::*;
 
-use super::messages::*;
-use super::tasks::Task;
-use crate::helpers::TaskStatus;
+// use super::messages::*;
+// use super::tasks::Task;
+// use crate::helpers::TaskStatus;
 use crate::schema::history;
